capturesync/requirements.txt
watchdog
Pillow
rich
customtkinter
face_recognition
numpy
opencv-python
